Add new files for Flatify component
This commit includes addition of new files and code elements for the Flatify component and its related functionalities. It also includes renaming and moving of the UserProfile.razor following the new organization of the files. The changes are made to complement the UI/UX upgrades and adjustments in the application.