Update color styles and add new functionalities to various components
Several components have been enhanced with the introduction of color styles and other features. For example, color styles have been added to the FlatButton, FlatSwitch, and FlatDateField components. Additional changes include adjustments to the tint and shade strength of the FlatThemeColor and updates to the Palette. New functionalities have been added to Table and Footer components. Adjustments to the FlatThemeProvider.razor also included style changes.