Add new features and refactor existing codes
The commit includes adding the 'ContentOptions' class and updating 'User' model to include 'ContentOptions'. New components for operations on Language like 'CreateLanguage' and 'EditLanguage' are introduced. Similarly, new pages: 'ViewWorld', 'Articles', and 'Manuals' are added. Refactoring includes removing the 'ManualsHome' page, and adjustments in 'FlatTextArea', 'RequestComponent', and 'User' model. In addition, updates are made to the project files and several other codes.