Update FlatButton styles and add OnBlur event
The default button width is now set to the 'Default' value instead of 'Medium'. In addition, CSS class addition for width and edge type now considers their respective default conditions. An OnBlur event handling has also been introduced to FlatButtonBase, and the FlatButton element is updated to sync with this added event.